Migrate to Ktor 2
This commit migrates the project to Ktor 2.

In Ktor 2, `JsonFeature` is deprecated and superseded by a client
`ContentNegotiation` plugin, which results in the core serialization logic
being shared between client and server:
https://ktor.io/docs/migrating-2.html#serialization-client Hence, as part of
this migration, the client and server modules are combined into a single module
which has no client-specific or server-specific dependencies.

Several other upgrades (Kotlin, Gradle, and Moshi) are included in this change, as they were directly or indirectly required for the Ktor 2 upgrade.

Closes #47
Closes #48